VNQ-10366: Added separate activities for testing Markers and Clustering. Added some coordinates to test Clustering.
