+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="ru.dublgis.dgismobile.sdktestapp">
+        package="ru.dublgis.dgismobile.sdktestapp">
 
     <application
             android:allowBackup="true"
@@ -9,23 +9,30 @@
             android:roundIcon="@mipmap/icon_round"
             android:supportsRtl="true"
 
-            android:usesCleartextTraffic="true"
+            android:theme="@style/AppTheme"
 
-            android:theme="@style/AppTheme">
+            android:usesCleartextTraffic="true">
 
         <activity
                 android:name=".MainActivity"
                 android:configChanges="orientation|keyboardHidden"
-                android:theme="@style/SplashTheme"
-        >
+                android:theme="@style/SplashTheme">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+                android:name=".MarkersActivity"
+                android:configChanges="orientation|keyboardHidden"
+                android:theme="@style/SplashTheme" />
+        <activity
+                android:name=".ClusterersActivity"
+                android:configChanges="orientation|keyboardHidden"
+                android:theme="@style/SplashTheme" />
 
     </application>
 
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
+    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
 
 </manifest>
