+++ app/src/main/res/layout/activity_main.xml
@@ -1,83 +1,28 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-<androidx.drawerlayout.widget.DrawerLayout
-        xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:app="http://schemas.android.com/apk/res-auto"
-        xmlns:tools="http://schemas.android.com/tools"
-        android:id="@+id/drawer_layout"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        tools:openDrawer="start">
+        android:layout_height="wrap_content"
+        android:padding="8dp">
 
-    <androidx.constraintlayout.widget.ConstraintLayout
-            xmlns:android="http://schemas.android.com/apk/res/android"
-            xmlns:tools="http://schemas.android.com/tools"
-            xmlns:app="http://schemas.android.com/apk/res-auto"
+    <Button
+            android:id="@+id/markers"
             android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            tools:context=".MainActivity">
+            android:layout_height="wrap_content"
+            android:layout_marginTop="32dp"
+            android:text="Markers"
+            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintRight_toRightOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
 
-        <FrameLayout
-                android:id="@+id/map_container"
-                android:layout_width="0dp"
-                android:layout_height="0dp" app:layout_constraintEnd_toEndOf="parent"
+    <Button
+            android:id="@+id/clusterers"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="8dp"
+            android:text="Clustering"
+            app:layout_constraintLeft_toLeftOf="parent"
+            app:layout_constraintRight_toRightOf="parent"
+            app:layout_constraintTop_toBottomOf="@id/markers" />
 
-                app:layout_constraintTop_toTopOf="parent" app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintBottom_toBottomOf="parent">
-
-            <fragment
-                    android:id="@+id/mapFragment"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:name="ru.dublgis.dgismobile.mapsdk.MapFragment"
-            />
-
-        </FrameLayout>
-
-        <ImageButton
-                android:id="@+id/zoom_in"
-                android:layout_width="@dimen/map_control"
-                android:layout_height="@dimen/map_control"
-                android:background="@drawable/background"
-                android:src="@drawable/zoom_in"
-                android:elevation="2dp"
-
-                android:layout_marginRight="8dp"
-                app:layout_constraintRight_toRightOf="parent"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintBottom_toTopOf="@id/zoom_out"
-                app:layout_constraintVertical_chainStyle="packed"
-            />
-
-        <ImageButton
-                android:id="@+id/zoom_out"
-                android:layout_width="@dimen/map_control"
-                android:layout_height="@dimen/map_control"
-                android:background="@drawable/background"
-                android:src="@drawable/zoom_out"
-                android:elevation="2dp"
-
-                android:layout_marginRight="8dp"
-                android:layout_marginTop="12dp"
-                app:layout_constraintRight_toRightOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/zoom_in"
-                app:layout_constraintBottom_toTopOf="@id/location"
-                />
-        <ImageButton
-                android:id="@+id/location"
-                android:layout_width="@dimen/map_control"
-                android:layout_height="@dimen/map_control"
-                android:background="@drawable/background"
-                android:src="@drawable/location"
-                android:elevation="2dp"
-
-                android:layout_marginRight="8dp"
-                android:layout_marginTop="12dp"
-                app:layout_constraintRight_toRightOf="parent"
-                app:layout_constraintTop_toBottomOf="@id/zoom_out"
-                app:layout_constraintBottom_toBottomOf="parent"
-                />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-</androidx.drawerlayout.widget.DrawerLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
