+++ app/src/main/res/layout/activity_map.xml
@@ -1,0 +1,83 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<androidx.drawerlayout.widget.DrawerLayout
+        xmlns:android="http://schemas.android.com/apk/res/android"
+        xmlns:app="http://schemas.android.com/apk/res-auto"
+        xmlns:tools="http://schemas.android.com/tools"
+        android:id="@+id/drawer_layout"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:openDrawer="start">
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+            xmlns:android="http://schemas.android.com/apk/res/android"
+            xmlns:tools="http://schemas.android.com/tools"
+            xmlns:app="http://schemas.android.com/apk/res-auto"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            tools:context=".MainActivity">
+
+        <FrameLayout
+                android:id="@+id/map_container"
+                android:layout_width="0dp"
+                android:layout_height="0dp" app:layout_constraintEnd_toEndOf="parent"
+
+                app:layout_constraintTop_toTopOf="parent" app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintBottom_toBottomOf="parent">
+
+            <fragment
+                    android:id="@+id/mapFragment"
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent"
+                    android:name="ru.dublgis.dgismobile.mapsdk.MapFragment"
+            />
+
+        </FrameLayout>
+
+        <ImageButton
+                android:id="@+id/zoom_in"
+                android:layout_width="@dimen/map_control"
+                android:layout_height="@dimen/map_control"
+                android:background="@drawable/background"
+                android:src="@drawable/zoom_in"
+                android:elevation="2dp"
+
+                android:layout_marginRight="8dp"
+                app:layout_constraintRight_toRightOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintBottom_toTopOf="@id/zoom_out"
+                app:layout_constraintVertical_chainStyle="packed"
+            />
+
+        <ImageButton
+                android:id="@+id/zoom_out"
+                android:layout_width="@dimen/map_control"
+                android:layout_height="@dimen/map_control"
+                android:background="@drawable/background"
+                android:src="@drawable/zoom_out"
+                android:elevation="2dp"
+
+                android:layout_marginRight="8dp"
+                android:layout_marginTop="12dp"
+                app:layout_constraintRight_toRightOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/zoom_in"
+                app:layout_constraintBottom_toTopOf="@id/location"
+                />
+        <ImageButton
+                android:id="@+id/location"
+                android:layout_width="@dimen/map_control"
+                android:layout_height="@dimen/map_control"
+                android:background="@drawable/background"
+                android:src="@drawable/location"
+                android:elevation="2dp"
+
+                android:layout_marginRight="8dp"
+                android:layout_marginTop="12dp"
+                app:layout_constraintRight_toRightOf="parent"
+                app:layout_constraintTop_toBottomOf="@id/zoom_out"
+                app:layout_constraintBottom_toBottomOf="parent"
+                />
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+</androidx.drawerlayout.widget.DrawerLayout>
